+++ tests/Grant/DeviceCodeGrantTest.php
@@ -20,7 +20,6 @@ use League\OAuth2\Server\Repositories\RefreshTokenRepositoryInterface;
 use League\OAuth2\Server\Repositories\ScopeRepositoryInterface;
 use LeagueTests\Stubs\AccessTokenEntity;
 use LeagueTests\Stubs\ClientEntity;
-use LeagueTests\Stubs\CryptTraitStub;
 use LeagueTests\Stubs\DeviceCodeEntity;
 use LeagueTests\Stubs\RefreshTokenEntity;
 use LeagueTests\Stubs\ScopeEntity;
@@ -28,22 +27,13 @@ use LeagueTests\Stubs\StubResponseType;
 use PHPUnit\Framework\TestCase;
 
 use function base64_encode;
-use function json_encode;
 use function random_bytes;
-use function time;
 use function uniqid;
 
 class DeviceCodeGrantTest extends TestCase
 {
     private const DEFAULT_SCOPE = 'basic';
     private const INTERVAL_RATE = 10;
-
-    protected CryptTraitStub $cryptStub;
-
-    public function setUp(): void
-    {
-        $this->cryptStub = new CryptTraitStub();
-    }
 
     public function testGetIdentifier(): void
     {
@@ -102,7 +92,6 @@ class DeviceCodeGrantTest extends TestCase
 
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setScopeRepository($scopeRepositoryMock);
 
         $request = (new ServerRequest())->withParsedBody([
@@ -147,7 +136,6 @@ class DeviceCodeGrantTest extends TestCase
 
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setScopeRepository($scopeRepositoryMock);
 
         $request = (new ServerRequest())->withParsedBody([
@@ -261,6 +249,7 @@ class DeviceCodeGrantTest extends TestCase
     public function testCompleteDeviceAuthorizationRequest(): void
     {
         $deviceCode = new DeviceCodeEntity();
+        $deviceCode->setIdentifier('deviceCodeEntityIdentifier');
         $deviceCode->setUserCode('foo');
 
         $deviceCodeRepository = $this->getMockBuilder(DeviceCodeRepositoryInterface::class)->getMock();
@@ -273,9 +262,7 @@ class DeviceCodeGrantTest extends TestCase
             'http://foo/bar',
         );
 
-        $grant->setEncryptionKey($this->cryptStub->getKey());
-
-        $grant->completeDeviceAuthorizationRequest($deviceCode->getUserCode(), 'userId', true);
+        $grant->completeDeviceAuthorizationRequest($deviceCode->getIdentifier(), 'userId', true);
 
         $this::assertEquals('userId', $deviceCode->getUserIdentifier());
     }
@@ -314,7 +301,7 @@ class DeviceCodeGrantTest extends TestCase
         $server->setDefaultScope(self::DEFAULT_SCOPE);
 
         $serverRequest = (new ServerRequest())->withParsedBody([
-           'client_id'     => 'foo',
+            'client_id'     => 'foo',
         ]);
 
         $deviceCodeGrant = new DeviceCodeGrant(
@@ -323,8 +310,6 @@ class DeviceCodeGrantTest extends TestCase
             new DateInterval('PT10M'),
             'http://foo/bar'
         );
-
-        $deviceCodeGrant->setEncryptionKey($this->cryptStub->getKey());
 
         $server->enableGrantType($deviceCodeGrant);
 
@@ -346,15 +331,8 @@ class DeviceCodeGrantTest extends TestCase
         $clientRepositoryMock->method('getClientEntity')->willReturn($client);
         $clientRepositoryMock->method('validateClient')->willReturn(true);
 
-        $accessTokenRepositoryMock = $this->getMockBuilder(AccessTokenRepositoryInterface::class)->getMock();
-        $accessToken = new AccessTokenEntity();
-        $accessToken->setClient($client);
-        $accessTokenRepositoryMock->method('getNewToken')->willReturn($accessToken);
-        $accessTokenRepositoryMock->method('persistNewAccessToken')->willReturnSelf();
-
-        $refreshTokenRepositoryMock = $this->getMockBuilder(RefreshTokenRepositoryInterface::class)->getMock();
-        $refreshTokenRepositoryMock->method('persistNewRefreshToken')->willReturnSelf();
-        $refreshTokenRepositoryMock->method('getNewRefreshToken')->willReturn(new RefreshTokenEntity());
+        $scope = new ScopeEntity();
+        $scope->setIdentifier('foo');
 
         $deviceCodeRepositoryMock = $this->getMockBuilder(DeviceCodeRepositoryInterface::class)->getMock();
         $deviceCodeEntity = new DeviceCodeEntity();
@@ -364,12 +342,28 @@ class DeviceCodeGrantTest extends TestCase
         $deviceCodeEntity->setUserCode('123456');
         $deviceCodeEntity->setExpiryDateTime(new DateTimeImmutable('+1 hour'));
         $deviceCodeEntity->setClient($client);
+        $deviceCodeEntity->addScope($scope);
 
-        $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCodeEntity);
+        $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')
+            ->with($deviceCodeEntity->getIdentifier())
+            ->willReturn($deviceCodeEntity);
 
-        $scope = new ScopeEntity();
+        $accessTokenEntity = new AccessTokenEntity();
+        $accessTokenEntity->setClient($client);
+        $accessTokenEntity->addScope($scope);
+
+        $accessTokenRepositoryMock = $this->getMockBuilder(AccessTokenRepositoryInterface::class)->getMock();
+        $accessTokenRepositoryMock->method('getNewToken')
+            ->with($client, $deviceCodeEntity->getScopes(), $deviceCodeEntity->getUserIdentifier())
+            ->willReturn($accessTokenEntity);
+        $accessTokenRepositoryMock->method('persistNewAccessToken')->willReturnSelf();
+
+        $refreshTokenRepositoryMock = $this->getMockBuilder(RefreshTokenRepositoryInterface::class)->getMock();
+        $refreshTokenRepositoryMock->method('persistNewRefreshToken')->willReturnSelf();
+        $refreshTokenRepositoryMock->method('getNewRefreshToken')->willReturn(new RefreshTokenEntity());
+
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -383,26 +377,13 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setAccessTokenRepository($accessTokenRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
-        $grant->completeDeviceAuthorizationRequest($deviceCodeEntity->getUserCode(), '1', true);
+        $grant->completeDeviceAuthorizationRequest($deviceCodeEntity->getIdentifier(), 'baz', true);
 
         $serverRequest = (new ServerRequest())->withParsedBody([
             'grant_type' => 'urn:ietf:params:oauth:grant-type:device_code',
-            'device_code'   => $this->cryptStub->doEncrypt(
-                json_encode(
-                    [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() + 3600,
-                        'client_id' => 'foo',
-                        'user_code' => '12345678',
-                        'scopes' => ['foo'],
-                        'verification_uri' => 'http://foo/bar',
-                    ],
-                    JSON_THROW_ON_ERROR
-                )
-            ),
+            'device_code'   => $deviceCodeEntity->getIdentifier(),
             'client_id'     => 'foo',
         ]);
 
@@ -410,6 +391,7 @@ class DeviceCodeGrantTest extends TestCase
         $grant->respondToAccessTokenRequest($serverRequest, $responseType, new DateInterval('PT5M'));
 
         $this::assertInstanceOf(RefreshTokenEntityInterface::class, $responseType->getRefreshToken());
+        $this::assertSame([$scope], $responseType->getAccessToken()->getScopes());
     }
 
     public function testRespondToRequestMissingClient(): void
@@ -432,19 +414,7 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setAccessTokenRepository($accessTokenRepositoryMock);
 
         $serverRequest = (new ServerRequest())->withQueryParams([
-            'device_code' => $this->cryptStub->doEncrypt(
-                json_encode(
-                    [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() + 3600,
-                        'client_id' => 'foo',
-                        'user_code' => '12345678',
-                        'scopes' => ['foo'],
-                        'verification_uri' => 'http://foo/bar',
-                    ],
-                    JSON_THROW_ON_ERROR
-                )
-            ),
+            'device_code' => uniqid(),
         ]);
 
         $responseType = new StubResponseType();
@@ -471,9 +441,8 @@ class DeviceCodeGrantTest extends TestCase
         $deviceCodeEntity->setUserIdentifier('baz');
         $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCodeEntity);
 
-        $scope = new ScopeEntity();
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -486,7 +455,6 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
         $serverRequest = (new ServerRequest())->withParsedBody([
@@ -520,9 +488,8 @@ class DeviceCodeGrantTest extends TestCase
         $deviceCodeEntity->setClient($client);
         $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCodeEntity);
 
-        $scope = new ScopeEntity();
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -535,24 +502,11 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
         $serverRequest = (new ServerRequest())->withParsedBody([
             'client_id'     => 'foo',
-            'device_code'   => $this->cryptStub->doEncrypt(
-                json_encode(
-                    [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() + 3600,
-                        'client_id' => 'foo',
-                        'user_code' => '12345678',
-                        'scopes' => ['foo'],
-                        'verification_uri' => 'http://foo/bar',
-                    ],
-                    JSON_THROW_ON_ERROR
-                )
-            ),
+            'device_code'   => uniqid(),
         ]);
 
         $responseType = new StubResponseType();
@@ -581,9 +535,8 @@ class DeviceCodeGrantTest extends TestCase
         $deviceCodeEntity->setClient($client);
         $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCodeEntity);
 
-        $scope = new ScopeEntity();
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -596,24 +549,11 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
         $serverRequest = (new ServerRequest())->withParsedBody([
             'client_id'     => 'foo',
-            'device_code'   => $this->cryptStub->doEncrypt(
-                json_encode(
-                    [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() + 3600,
-                        'client_id' => 'foo',
-                        'user_code' => '12345678',
-                        'scopes' => ['foo'],
-                        'verification_uri' => 'http://foo/bar',
-                    ],
-                    JSON_THROW_ON_ERROR
-                )
-            ),
+            'device_code'   => uniqid(),
         ]);
 
         $responseType = new StubResponseType();
@@ -642,9 +582,8 @@ class DeviceCodeGrantTest extends TestCase
         $deviceCodeEntity->setClient($client);
         $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCodeEntity);
 
-        $scope = new ScopeEntity();
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -657,24 +596,11 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
         $serverRequest = (new ServerRequest())->withParsedBody([
             'client_id'     => 'foo',
-            'device_code'   => $this->cryptStub->doEncrypt(
-                json_encode(
-                    [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() - 3600,
-                        'client_id' => 'foo',
-                        'user_code' => '12345678',
-                        'scopes' => ['foo'],
-                        'verification_uri' => 'http://foo/bar',
-                    ],
-                    JSON_THROW_ON_ERROR
-                )
-            ),
+            'device_code'   => uniqid(),
         ]);
 
         $responseType = new StubResponseType();
@@ -713,7 +639,6 @@ class DeviceCodeGrantTest extends TestCase
 
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setIntervalVisibility(true);
 
@@ -748,14 +673,14 @@ class DeviceCodeGrantTest extends TestCase
 
         $deviceCode = new DeviceCodeEntity();
 
+        $deviceCode->setIdentifier('deviceCodeEntityIdentifier');
         $deviceCode->setExpiryDateTime(new DateTimeImmutable('+1 hour'));
         $deviceCode->setClient($client);
         $deviceCode->setUserCode('12345678');
         $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCode);
 
-        $scope = new ScopeEntity();
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -768,26 +693,13 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
-        $grant->completeDeviceAuthorizationRequest($deviceCode->getUserCode(), '1', false);
+        $grant->completeDeviceAuthorizationRequest($deviceCode->getIdentifier(), '1', false);
 
         $serverRequest = (new ServerRequest())->withParsedBody([
-                'client_id'     => 'foo',
-                'device_code'   => $this->cryptStub->doEncrypt(
-                    json_encode(
-                        [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() + 3600,
-                        'client_id' => 'foo',
-                        'user_code' => '12345678',
-                        'scopes' => ['foo'],
-                        'verification_uri' => 'http://foo/bar',
-                        ],
-                        JSON_THROW_ON_ERROR
-                    ),
-                ),
+            'client_id'     => 'foo',
+            'device_code'   => $deviceCode->getIdentifier(),
         ]);
 
         $responseType = new StubResponseType();